bugfix for rotating hobbit
git-svn-id: https://repo.acin.tuwien.ac.at/v4r/trunk/ros/hobbit_hydro_src@10757 a44b1c4d-ba22-49c9-afbd-3e401e1c4b3e

+++ hobbit_smach/src/hobbit_smach/hobbit_move_import.py
@@ -18,7 +18,7 @@ from hobbit_msgs.srv import GetCoordinates, GetCoordinatesRequest, GetName, \
     SwitchVision, SwitchVisionRequest, SendValueRequest, SendValue
 from move_base_msgs.msg import MoveBaseAction
 from hobbit_msgs.msg import MiraDockingAction, MiraDockingGoal
-from std_msgs.msg import String
+from std_msgs.msg import String, Float32
 from hobbit_user_interaction import HobbitMMUI, HobbitEmotions
 from uashh_smach.util import SleepState, WaitForMsgState
 from actionlib import SimpleActionClient
@@ -280,7 +280,7 @@ def move_discrete(in_motion=None, in_value=None):
                 ServiceState(
                     '/apply_rotation',
                     SendValue,
-                    request=SendValueRequest(value=in_value),
+                    request=SendValueRequest(value=Float32(in_value)),
                     response_cb=send_value_resp_cb
                 ),
                 transitions={'succeeded': 'SET_DOCK_STATE_VAR',
